backend/requirements.txt
fastapi
langchain
pydantic[email]
langchain-community
langchain-google-genai
langchain-mongodb
langchain-text-splitters    
pymupdf
pymongo
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
supabase
gotrue
python-multipart
pytest
pytest-asyncio
httpx
structlog
langchain-huggingface